salt: Move kubeconfig validate method to a dedicated exec mod
This way this method can be called by anything to check
if the kubeconfig file needs to be regenerated.

The purpose is to be able to call it from a custom
beacon to check whether embedded certificates are
expired or not.

Refs: #1887
